+++ doc/whats-new.rst
@@ -83,8 +83,8 @@ New Features
 - Added ability to save ``DataArray`` objects directly to Zarr using :py:meth:`~xarray.DataArray.to_zarr`.
   (:issue:`7692`, :pull:`7693`) .
   By `Joe Hamman <https://github.com/jhamman>`_.
-- Added `numpy_data` keyword argument to both :py:meth:`xarray.Dataset.to_dict` and
-  :py:meth:`xarray.DataArray.to_dict` to return data as numpy objects instead of native Python objects.
+- Keyword argument `data='array'` to both :py:meth:`xarray.Dataset.to_dict` and
+  :py:meth:`xarray.DataArray.to_dict` will now return data as numpy.ndarrays. Python lists are returned for `data='list'` or `data=True`. Supplying `data=False` only returns the schema without data.
   (:issue:`1599`, :pull:`7739`) .
   By `James McCreight <https://github.com/jmccreight>`_.
 - Support `pandas>=2.0` (:pull:`7724`)